Don't reset ExternalPointerSlot in JSSynchronizationPrimitive::SetWaiterQueueHead
Currently, ExternalPointerSlots are effectively const-after-init and so
we should not reset them to the null ExternalPointerHandle as that could
for example lead to DCHECK failures during compaction (though it's
probably fine in practice). Instead, we should just set the value in the
table entry to nullptr. In the future, we could consider relaxing this
requirement and allowing external pointer handles to be arbitrarily
modified. This would mostlly require changing the table compaction
algorithm and probably adding a few more write barriers (for which the
infrastructure already exists).

Bug: 378066790
Change-Id: I3dfc8bdeee39b33886e37c6eb6363ecf1edea765
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6020068
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97196}