[sandbox] Rename OLD_TO_CODE to TRUSTED_TO_CODE...
... and ensure that the set is always located outside of the sandbox.

Similar to the TRUSTED_TO_TRUSTED remembered set, this remembered set
should only be used in the trusted space (outside the sandbox) as it
contains references from Code objects (in trusted space) to
InstructionStream objects (in code space). As such, it is assumed that
the compressed pointers in it are offsets into the code space. However,
an attacker could create such a remembered set inside the sandbox, in
which case we might end up corrupting memory inside the (trusted) code
space during mark-compact GC. As such, this CL now adds a SBXCHECK that
the remembered set is not inside the sandbox. In addition, we also need
to make sure that we're not inserting invalid pointers into (legitimate)
TRUSTED_TO_CODE or TRUSTED_TO_TRUSTED remembered sets, which currently
requires another SBXCHECK (but see also issue 377724745).

Change-Id: I07d17bd92b7b58668fe07e4a9b149bc8b6356db6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6023179
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97185}