[maglev] Emit representation-specific checks in BuildCheckSmi
Maglev automatically converts node inputs based on their
representation. For example, it converts CheckSmi input
to Tagged if the input is Int32.

However, in BuildCheckSmi, before emitting CheckSmi, we
already ensure that the input representation fits within an
Smi and emit an UnsafeSmiTagInt32.

This is naturally wrong for any Int32.

In reality, BuildCheckSmi is meant to check whether the
tagged variant of the node is a Smi. Ie, for nodes with
different representation, it should check if it fits within
an Smi. In this CL, we modify BuildCheckSmi to emit a check
for each input representation directly, simplifying the logic
and avoiding redundant conversions.

Fixed: 378836891
Change-Id: I6c7e222b9df1a6c52c4cfb62097ea86f961f7c7a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6030493
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97239}