[scopes] Reattach preexisting scope infos to recreated SFI
Previously it was possible that we compiled an SFI, causing it to
acquire a ScopeInfo also held alive by inner functions. That SFI could
die if the outer function's bytecode died. On recreation the scope
info still held alive by inner code wouldn't eagerly be reattached. It
was only reattached on compilation.

That would have been fine if it weren't for complications coming from
background merging of code caches with isolate caches (so while the
main thread could also be compiling some of the same code). A code
cache from disk that's being background merged could have spotty
coverage of the code. So can the isolate cache. E.g. for this SFI both
the isolate cache and the disk cache could have the SFI but without
code or scope info. The on-disk cache could have live inner scope
holding on to the scope info of that function. So far this would be
fine as later compilation of the function would pick up the inner
info. However, if the main thread, before finalization, were to
compile the SFI, it would recreate the scope info (the main thread
doesn't know anything about the inner code that's being deserialized).
During merge finalization we didn't realise this happened, and ended
with 2 scope infos for the same function.

This CL ensures that we always reattach existing scope infos to
recreated SFIs. If there's live inner code to the SFI, we'll pick up
the scope info from there on SFI creation rather than on compilation.

This means that if the main-thread version of the SFI has a live scope
info, we'll either pick it up from the SFI directly during background
merging (this could have been missed previously if it was only
reachable through live inner code), or we'll find it during foreground
merge completion if new inner code appeared on the main thread. If
neither of those cases happened; and if the main thread SFI is
uncompiled; we'll be able to ignore the main thread state as no
main-thread code cares about it any longer. (If the main thread code
is compiled we'll simply keep the state as is.)


Change-Id: I1c6a0908c565011ded5969d5b95e029b8ad38006
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6030933
Auto-Submit: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97243}