[handles] Replace PatchValue with SetValue for direct handles
PatchValue for indirect handles serves the purpose of updating the
handle's value in such a way that all other indirect handles with the
same location are affected. This is not possible with direct handles,
where the current implementation of this method is semantically
equivalent to simple handle assignment.

This CL removes the method PatchValue for direct handles, as its use
during migration from indirect to direct handles would introduce subtle
bugs that would be difficult to catch. It replaces it with a SetValue
method, which serves as in-place handle assignment. To support the
configuration with "fake" direct handles that is currently used in
production, SetValue for such handles is implemented via PatchValue
(and therefore affects all indirect handles with the same location),
in order to avoid performance regressions.

The CL also aligns a few differences in the implementation of "fake"
direct handles and maybe-handles. It introduces the auxiliary function
direct_handle, for ensuring that a handle is converted to direct.

Bug: 42203211
Change-Id: Ie72d2fba7cdfdb47db094cbedea5613c58d375a2
Cq-Include-Trybots: luci.v8.try:v8_linux64_css_dbg
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6018328
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97302}