[heap] Iterate shared heap object cache also for worker isolates
This condition was still assuming that the existence of a shared heap
implies --shared-string-table. However, with the --shared-heap flag,
there is a shared heap but client isolates still have separate string
tables and therefore the shared heap object cache needs to be part of
the root set.

The implication from Isolate::OwnsStringTables() to
IterateSharedHeapObjectCache() might be a bit surprising but will
become clearer once it gets renamed to IterateSharedStringTableCache()
in a follow-up.

Bug: 372493838, 379800425
Change-Id: I3a39f77bb01eb5e268b7315c3179b269328eea91
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6035115
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97287}