[maglev] Refactor predecessor_count accounting
Enforce a new invariant if a merge state exists, then
`MaglevGraphBuilder::predecessor_count_[offset]` should be in sync with
`merge_states[offset]->predecessor_count_`.

This should give us more confidence that the accounting is correct and
we are in a consistent state.

This might fix some dchecks which triggered when we check the
predecessor count on the builder -- which might be out of sync with the
actual count. Or, the additional dchecks might get us closer to the
source of these crashes.

Bug: 378903450
Change-Id: I4ac93849a650a58d949c24cede7f1abad9652955
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6032550
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97263}