[wasm][jspi] Add missing builtins to wasm-builtin-list.h
Add WasmResume and WasmReject to the wasm builtin list. Adding them to
this list sets the is_wasm assembler option, which turns on the
stack-switching variant of the CEntry builtin. In particular this
ensures that that the WasmReject builtin switches to the central stack
before calling the Throw runtime function.

Drive-by: remove dead code related to inactive stacks in
heap::base::Stack.

R=clemensb@chromium.org,nikolaos@chromium.org

Fixed: 42203436
Change-Id: I0e3ca1ab6cc8b70df0a9e505101d0dd1977b7c40
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6035509
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97280}