[sandbox] Add some hardening around parameter counts in src/compiler
The JSDispatchTable is used to ensure that callee and caller agree on
the signature during a JS function call. This works by making sure that
the parameter count of a dispatch entry always matches the parameter
count of the code installed in it. This in turn relies on the parameter
count reported by the Code object. As such, during compilation we need
to make sure that the parameter_count reported by the produced Code
object matches the parameter count used/expected by the generated code.
Otherwise, we may still end up with (effectively) signature mismatches.
This CL adds a few defense-in-depth CHECKs for that purpose.

Bug: 40931165
Change-Id: I776258ba3ac42884f6363134a07e5f8389fec2e4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6031884
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97285}