[heap] Use IsMarkedOrAlwaysLive() helper in object-stats.cc
Use IsMarkedOrAlwaysLive() instead of checking markbits manually. This
should now handle our new "black allocation" scheme correctly as well.

This CL should also fix a data race with --shared-string-table between
the main and worker isolates. In this race the worker isolate collects
object stats and checks the mark bit of a shared string. Concurrently
the main thread sweeps the shared space page of that shared string.

I was not able to reproduce this race locally but it should be
fixed because by using IsMarkedOrAlwaysLive() we are not loading
markbits for shared objects on worker isolates anymore.

Bug: 379805756
Change-Id: I1cc1625521c5d699eacff4f9df258781c8f065c8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6037996
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97306}