[heap] Revise ResourceConstraints::ConfigureDefaults and tests
`ResourceConstraints::ResourceConstraints` and
`ResourceConstraints::ConfigureDefaultsFromHeapSize` are only used
from WebView and not from the main path. On ToT they return half
intended the max new space size. This CL fixes that so that the
intended max size based on the flags is always returned.

This CL also revises the tests for these methods. Previously the tests
used mostly hardcoded values which made the test fragile and hard to
maintain. Everytime we changed flag defaults, these tests would break
with unexpected values. The tests should be more robust and rely mostly
on constants (that are implicitly updated when flags change).

Bug: 380011691
Change-Id: I8ded510d3c5dc7914421b6091769b55dee93da09
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6038843
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97498}