[sandbox][wasm] Always copy Wasm wire bytes
With the sandbox attacker model, it is possible also for non-shared
ArrayBuffers to be concurrently modified. To better protect against
that, this CL changes the Wasm pipeline to always create a copy of the
wire bytes. For that, it changes the main entrypoints into the
WasmEngine (SyncCompile, AsyncCompile) to require a base::OwnedVector
for the wire bytes instead of an (unowned) ModuleWireBytes. Besides
enforcing that on-heap data is copied out of the heap, this also allows
the engine to avoid copying the bytes in GetOrCompileNewNativeModule. As
such, in practice this change should mostly not lead to additional
copies as asynchronous and streaming compilation were already copying
the data initially, and synchronous compilation was often copying the
data later on anyway.

Drive-By: rename OwnedVector<T>::Of to OwnedVector<T>::NewByCopying to
be more explicit about making a copy and migrate existing users to the
new and more convenient base::OwnedCopyOf(...) functions.

Bug: 376071292
Change-Id: I0b3c56d56c744bbf64f43a8b5a142ee2a864afed
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6037693
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97343}