[csa] Refactor CodeAssemblerCompilationJob
This CL refactors CodeAssemblerCompilationJob into a Turbofan subclass
and a Turboshaft subclass, as well as put the base class into its own
header. This sets it up to be used in tests so that
Pipeline::GenerateCodeForCodeStub can be removed in favor of always
using the job.

Bug: 342692713
Change-Id: I80586ea87f60cfa31c0d40c2e4571857afd1feb1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6039284
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97357}