[sandbox] Check stack size on OSR entry
This ensures that it is not possible to corrupt the stack height
by OSR'ing into incompatible code.

The check consists of two parts. The OSR-out side ensures that the OSR
code object agrees on the number of arguments. The OSR-in side ensures
that the stack height matches.

Bug: 354724082, 374812612
Change-Id: If7842d50fd5131ebc6b04b2bf13fcaf1b83e3b43
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5999476
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97335}