[etw]  Enable ETW tracing by default via custom filter
Add a new flag --enable-etw-by-custom-filter-only to control ETW
tracing. Unlike the existing flag --enable-etw-stack-walking this new
flag would be enabled by default, but it should only start an ETW
tracing session on Windows when the WPR (Windows Performance Recorder)
profile is configured with a custom filter.

We already support ETW custom filters to only enable ETW tracing for
isolates that are running in a document that navigates to one of a set
of specified URLs, in order to minimize the cost of ETW tracing. The
purpose of this new flag is to facilitate the automation of ETW
tracing sessions in user machines, without needing to launch the
browser with command-line flags.

A CustomFilter added to an EventProvider in a WPR profile has the
following format:

  <EventProvider Id="Edge_Stable" Name="57277741-3638-4A4B-BDBA-0AC6E45DA56C" Level="5">
    <CustomFilter Type="0x80000000" Value="AQABAAAAAAA..." />
  </EventProvider>

Where:
- Name="57277741-3638-4A4B-BDBA-0AC6E45DA56C" is the GUID of the V8
    ETW provider, (see src/libplatform/etw/etw-provider-win.h),
- Type="0x80000000" is EVENT_FILTER_TYPE_SCHEMATIZED,
- Value="AQABAAAAAA..." is a base64-encoded byte array that is
    base64-decoded by Windows and passed to the ETW enable callback in
    the 'PEVENT_FILTER_DESCRIPTOR FilterData' argument; see:
https://learn.microsoft.com/en-us/windows/win32/api/evntprov/ns-evntprov-event_filter_descriptor.

This array contains a struct EVENT_FILTER_HEADER followed by a
variable length payload, which is a string in JSON format:
  {
    "version": "2.0",
    "filtered_urls": [
        "https:\/\/.*\.chromium\.org\/.*", "https://v8.dev/.*", "..."
    ],
    "trace_interpreter_frames": false
  }

Where:
- filtered_urls is a list of regular expressions that should match the
    document URL in order to enable ETW tracing:
- trace_interpreter_frames is a new flag that controls whether
    interpreted functions should also be logged in a ETW tracing session
    initiated with this custom filter. The tracing of interpreter stack
    frames is slightly more expensive because it requires generating a
    distinct InterpreterEntryTrampoline for each function run by the
    interpreter.

Bug: 1346064442
Change-Id: I41e2001d4f76f701d23cf188697a025eaa344ad4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6027488
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#97527}