[json] Fast-path for stringify
Create a JSON.stringiy fast-path without side-effects/GCs.
We use a segmented, zone allocated  buffer to avoid copies as much as
possible.

The whole class is templatized on the buffers character type to avoid
runtime checks as much as possible. Since we already check instance
types exhaustively, as we don't deal with indirect strings that could
cause GCs, detecting encoding changes is fairly cheap.

Instead of the previous implementation, this version is iterative.
This has 2 main advantages:
1) Avoid stack overflow checks.
2) Easy to resume after detecting encoding changes. The goal is to also
support fast resume in the slow-path in the future.

Bug: 380044242

Change-Id: I5bdfc04f02aecdf744758c4af193817898ff7055
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6011806
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97381}