Fix another dead lock with transition arrays
https://chromium-review.googlesource.com/c/v8/v8/+/4062039 fixed a
deadlock, but there's another similar one (a subsequent allocation).
(See that CL for the description of the deadlock.)

This fix is based on the observation that the main thread is the only
writer, so it can read without holding the lock. This CL moves the
lock until the last moment, right before we'll write. Thus, all the
allocations now happen without holding the lock, removing the
deadlock.

Bug: 377885338
Change-Id: I7fcc7f83504959d8155b8864ca3e78a0d3deed82
Fixed: 377885338
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6048533
Commit-Queue: Marja Hölttä <marja@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97391}