[runtime] Don't compare negative zero float (part II)
This is a follow-up of https://crrev.com/c/6030009

Context stores usually use the code in contexts.tq, but if
the store is inside an eval, it falls back to the runtime
and goes into contexts.cc instead. A missing check was
accepting 0 and -0 as equals.

Fixed: 378891973
Change-Id: Ic287c120c2bbc2bca090ffea3f7bd7afc0a2561b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6049258
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97400}