[test] Avoid compiler flake with array allocation
To test compiler optimizations in mjsunit tests, it is common to use
the combination of `PrepareFunctionForOptimization` followed by
`OptimizeFunctionOnNextCall`. Between these two calls, the
to-be-optimized function has to be called to collect feedback. Typically,
a single call to the function is sufficient. However, when an array gets
created in the function, then a single call is not sufficient to
determine the type of the elements in the array. In that case, at least
two calls are needed. Otherwise, incomplete feedback is collected, and
the optimized function will deopt immediately.

With this CL, the to-be-optimized function gets called twice in a helper
function for testing fast api calls, which is used by some functions
where the array is created in the function.

Bug: 380898918
Change-Id: I44c9b89eead16ee55021fdc7ef6394b5c1622348
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6049418
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Auto-Submit: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97401}