[sandbox] Initialize HeapObjectData::map_ to nullptr initially
In the HeapObjectData constructor, the to-be-constructed data may
already be in the RefsMap (essentially a cache). Previously, we were
initializing the map_ field to the result of a GetOrCreateData lookup
for the map object. This, however, can end up loading an object from
that same cache. As such, if the object graph is manipulated (e.g. via
in-sandbox memory corruption) such that we end up accessing ourselves,
then we'll get a partially-initialized object back. With this CL, we now
set map_ to nullptr initially to force a safe crash in that case.

Bug: 379985068
Change-Id: Icf551648c8f22d957fc6cb67b0a2aad71be7c09e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6048888
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97420}