[sandbox] Add regression test for crbug.com/379774687
This is a follow-up for https://crrev.com/c/6035110 that adds a
regression test that is able to trigger issues due to concurrent
modifications of JS source code while it is being parsed. The testcase
is effectively a mini-fuzzer for these issues and requires --random-seed
to trigger the issue reliably. This way the test case is fairly robust
and doesn't rely on a specific heap layout, and it is easily picked up
by "real" fuzzers for further mutations.

Bug: 379774687
Change-Id: I95d1991c997354a8afc5c299475862aa68fa577d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6048825
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97449}