[tiering] Fix another dcheck with --always-turbofan
When the testing flag `--always-turbofan` is on we can end up starting
a maglev compile job while the function is already turbofan compiled.

The reason is that if the function is marked for optimization but
not bytecode compiled, then the bytecode compilation will also trigger
a turbofan compilation.

Fixed: 381199183
Change-Id: I2c45c5f42b410375eb3de0f132c055bab87e3797
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6055481
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97511}