[leaptiering][ia32] Port to ia32
Port leaptiering to ia32. Since this is the first 32bit port we need
to change some details in the dispatch table to play nice with an
underlying entity table without continuous memory allocation.

To simplify the implementation we keep the entry size the same as on
64 bits for now.

Bug: 42204201

Change-Id: I121ba521720f103d6f9a550b5015d82e1e73bcba
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6055123
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97520}