[wasm][cleanup] De-templatize GetTypeForFunction
Most users pass in module-relative signatures; for the only
exception to that (WasmJSFunction), there is no difference
to canonicalized signatures anyway. And for pretty-printing
types, we want module-relative indices. So we can go back to
having only a single variant of GetTypeForFunction which
takes module-relative signatures as input.
This is a followup to crrev.com/c/5920449.

Change-Id: I4638840d610f90819f7ec27d0b10f6df2524ee16
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6064041
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97526}