[wasm] Reduce Zone space waste in type canonicalizer
Reset the Zone in case the copied type is not used because an identical
type was found.

To simplify things, this CL also inlines the
`AddRecursiveGroup(CanonicalType)` method because in the case of adding
a single signature we checked before that it does not exist, so a lot of
code is unnecessary for that case. In particular, we do not need the
Zone snapshot.

R=jkummerow@chromium.org

Bug: 381887331
Change-Id: I2871f6de8f72640d61b29963c2e00e0ea84c0431
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6066318
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97573}