[leaptiering] Remove separate context specialized handle
Remove the possibility for a JSFunction and its FeedbackCell to have
different dispatch handles (which allows that JSFunction to be context
specialized despite the feedback cell having many closures).

Instead, force them to have the same dispatch handle, and clear the
context specialized code on a one-to-many feedback cell transition. To
keep the first closure reasonably optimized when this happens, also
schedule an eager re-compilation for that dispatch handle, back to the
tier its function was in.

Change-Id: I7adf518a28b9413af762132d409471e1b6bdf3f5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6055559
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97522}