[wasm] Introduce flag to print off-heap memory consumption
This adds the --print-wasm-offheap-memory-size flag, which prints the
estimated off-heap memory consumption of NativeModules and the overall
WasmEngine when they get deallocated. When a module gets deallocated we
also print the current size of the WasmEngine, because that's often the
high watermark, and it will go down when deallocating / deregistering
the NativeModule from the WasmEngine.

Regular expressions in perf tests will use the first output that
matches, so this will use the higher number for the engine then.

R=jkummerow@chromium.org

Bug: 381996072
Change-Id: I135add13bbd4ded6154af95bc6145767070747ae
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6069988
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97539}