[turboshaft][wasm] Fix stale reference in array.fill and array.new
When calling to C++ for array.fill (if the length is more than 16
elements), Turboshaft passed the value in a 64 bit stack slot.
For reference values on 32 bits and on pointer-compressed builds, this
required converting a 32 bit tagged value into a 64 bit stack slot.
The emitted __ ChangeInt32ToInt64() makes the value untagged.
When performing GVN, these zero-extensions can be de-duplicated.
When spilled, it will be spilled in an untagged stack slot, e.g.
when calling into an allocation that can trigger a GC, making the
stack slot ignored by the GC and therefore stale afterwards.

This CL instead emits a stack slot that aligns with the input type
removing the need for any conversions (like zero-extensions) for all
types and furthermore also allowing us to mark the stack slot for the
C call as tagged. (These will not be GVN-ed and array.fill doesn't
allocate, so this isn't strictly required.)

Fixed: 381332096
Change-Id: If20a9a97ca8b75db4b7d533acb698efe859a94df
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6070000
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97552}