[wasm][fuzzer] Move Wasm fuzzer code to separate directory
This allows us to define a buganizer component for that directory via
DIR_METADATA, and also generally improves code organization.

As part of this move, we also rename the "wasm" fuzzer to "module"
fuzzer. Note that "async" and "streaming" also fuzz whole modules, but
there it is more implied by the existing name that they work on whole
modules.

R=jkummerow@chromium.org, machenbach@chromium.org

Bug: 381816415
Change-Id: Ic8422e78c57d715980e4da999f0c48dbb6f71044
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6069993
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97580}