[base] Support symbolized backtraces in signal handler
Previously, our signal handler would not attempt to symbolize backtraces
as that was not async signal safe. As such, backtraces simply consisted
of a list of addresses. With this CL we now use backtrace_symbols_fd in
that case which prints directly to the given fd and is suitable for this
use case (from a signal handler). This isn't perfect as we don't control
the format of the output and because the function names will be mangled,
but at least it's better than raw addresses.

Change-Id: I3dabecaac516f0388cf08901c0419ba2055f4965
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6070003
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97568}