[sandbox] Ensure BackingStore::buffer_start always points into the sandbox
An ArrayBuffer's buffer pointer must always point into the sandbox (e.g.
because we'll reference it via an offset from the start of the sandbox).
When the buffer is empty, we use a special EmptyBackingStoreBuffer()
constant which points at the end of the sandbox, right before a huge
guard region area. This way, any access to the "null" value will result
in a harmless crash, even with the maximum possible index. However, we
should also use the EmptyBackingStoreBuffer() when returning the backing
buffer via the V8 API. Otherwise, if we return nullptr, a user of the
API might compute `nullptr + array->ByteOffset()` which could result in
an address in the first 32GB of the address space. This CL therefore
changes the BackingStore::buffer_start() accessor to return the
EmptyBackingStoreBuffer() constant if the buffer is empty.

Bug: 381999810
Change-Id: I5f1d909ed2cb458d406bbe133f12be1167141e56
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6072811
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97569}