[json] Ensure no toJSON on JSArray prototype chain
In addition to checking that a JSArray's prototype is the initial
JSArray prototype and it doesn't have interesting properties, also
check that the JSArray prototype's prototype is the initial JSObject
prototype without interesting properties. This is necessary to
ensure that there is no toJSON on the whole prototype chain for arrays.

Bug: 380044242
Change-Id: I663a72f43794f6bc808ee62d42bca24996cd4645
Fixed: 382326128
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6074631
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97576}