[leaptiering] Deoptimize code by patching the dispatch table
Before this CL every optimized code started with a check if its corresponding code object is marked for deoptimization. We change this to instead eagerly patch the dispatch table ensuring we never actually end up calling into a deoptimized code object.

To do so it is necessary to add a new back-reference to the code object.

Drive-By: Remove UpdateContextSpecializedCode and friends and instead have a JSFunction::UpdateOptimizedCode which deals with context specialization as well as setting the back-reference for deoptimization.

Bug: 42204201
Change-Id: Ie00e3d3932039a4b5204c805f3058716fbdb49b7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6070381
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97577}