[deoptimizer][wasm] Fix GC crash in lazy deopt of JS-to-Wasm wrapper
For inlined js-to-wasm-wrappers it can happen that a JS function called
from wasm triggers a lazy deoptimization for the JS code the wrapper
returns to. In that case the return value (e.g. a Word64 for an i64
result in Wasm) has to be materialized lazily (the JS BigInt).

This lazy materialization might trigger GCs. During such a GC all
deoptimized frames are visited including the ContinuationFrame.
We should not push the not-yet-materialized value onto that frame as it
would be misinterpreted as a tagged stack value by the GC.

Fixed: chromium:381458039
Change-Id: I49a7d8fb342d052d32b0879de9a8c4084c7dafef
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6074513
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97598}