Fix TrapHandler tests for CET
When CET is enabled in Windows, we can't set the instruction pointer to
any address in the trap handler. If we set the instruction pointer to a
generated code, the process crashes with FAST_FAIL_SET_CONTEXT_DENIED.

For unit tests recovering the instruction pointer to a generated code,
this CL adds a trampoline function written in C++ and available at
compile time which jumps to the generated landing pad. Also, updates
the default recovery code to a C++ function.

Bug: 42201233
Change-Id: I4e0c8f7ac3dd32d52f197e0707d457cd75cccc04
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6072116
Commit-Queue: Choongwoo Han <choongwoo.han@microsoft.com>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97688}