[wasm] make DispatchTable and WasmJSFunctionData own wrapper CPT entries
The DispatchTable now tracks which entries are wrappers (including the
generic wrapper) and owns the WasmCodePointerTable entry for these.
We then re-use the entry for generic and compiled wrappers.
The WasmJSFunctionData works the same, except that it always has a
wrapper.

In a future CL, we can move the ownership to the
WasmImportWrapperCache (with reference counting), which will allow us
to tier up all wrappers with the same signature at the same time.

Change-Id: I143ccde6dbcca8e5b2167847d1e5883b28f8eb7f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6059859
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#97628}